requirements.txt
fastapi
uvicorn
openai
pandas
numpy
matplotlib
seaborn
scikit-learn
duckdb
requests
beautifulsoup4
python-multipart
python-dotenv
google-generativeai
# RAG and Vector Database
chromadb
sentence-transformers
langchain
langchain-community
langchain-openai
# Memory and Learning
redis
pickle-mixin
# Additional utilities
tiktoken
nltk